improvement(Atualização de Estado e Testes): Altera a lógica de atualizar os valores na tela em relação ao estado. Altera o tipo de dado do price no backend para Decimal(4,2) para se adaptar aos requerimentos do teste.

+++ back-end/src/database/models/product.js
@@ -7,7 +7,7 @@ module.exports = (sequelize, DataTypes) => {
       autoIncrement: true,
     },
     name: DataTypes.STRING,
-    price: DataTypes.FLOAT,
+    price: DataTypes.DECIMAL(4,2),
     urlImage: DataTypes.STRING,
   },{
     tableName: 'products',
